\name{BLSALLFOOD}
\docType{data}
\alias{BLSALLFOOD}
\title{BLSALLFOOD Data}
\description{
The monthly time series of the number of workers engaged in food industries in
the United States (January 1967 - December 1979).
}
\usage{
data(BLSALLFOOD)
}
\format{
A time series of 156 observations.
}
\source{
The data were obtained from the United States Bureau of Labor Statistics (BLS).
}
\keyword{datasets}