\name{Haibara}
\docType{data}
\alias{Haibara}
\title{Haibara Data}
\description{
A bivariate time series of the groundwater level and the atmospheric pressure
 that were observed at 10-minuite intervals at the Haibara observatory of the
 Tokai region, Japan.
}
\usage{
data(Haibara)
}
\format{
A data frame with 400 observations on the following 2 variables.
  \tabular{rl}{
  [, 1] \tab Groundwater level\cr
  [, 2] \tab Atmospheric pressure
  }
}
\source{
The data were offered by Dr. M. Takahashi and Dr. N. Matsumoto of National
 Institute of Advanced Industrial Science and Technology.
}
\keyword{datasets}
