\name{HAKUSAN}
\docType{data}
\alias{HAKUSAN}
\title{Ship's Navigation Data}
\description{
  A multivariate time series of a ship's yaw rate, rolling, pitching and rudder
  angles which were recorded every second while navigating across the Pacific
  Ocean.
}
\usage{
data(HAKUSAN)
}
\format{
  A data frame with 1000 observations on the following 4 variables.
  \tabular{rll}{
  [, 1] \tab YawRate \tab yaw rate\cr
  [, 2] \tab Rolling \tab rolling\cr
  [, 3] \tab Pitching\tab pitching\cr
  [, 4] \tab Rudder  \tab rudder angle
  }
}
\source{
  The data were offered by Prof. K. Ohtsu of Tokyo University of Marine Science
  and Technology.
}
\keyword{datasets}
