\name{MYE1F}
\docType{data}
\alias{MYE1F}
\title{Seismic Data}
\description{
The time series of East-West components of seismic waves, recorded every 0.02
seconds.
}
\usage{
data(MYE1F)
}
\format{
A time series of 2600 observations.
}
\source{
Takanami, T. (1991), "ISM data 43-3-01: Seismograms of foreshocks of 1982
 Urakawa-Oki earthquake", Ann. Inst. Statist. Math., 43, 605.
}
\keyword{datasets}
