\name{Nikkei225}
\docType{data}
\alias{Nikkei225}
\title{Nikkei225}
\description{
A daily closing values of the Japanese stock price index, Nikkei225, quoted from
 January 4, 1988, to December 30, 1993.
}
\usage{
data(Nikkei225)
}
\format{
A time series of 1480 observations.
}
\source{
\url{https://indexes.nikkei.co.jp/nkave/archives/data}
}
\keyword{datasets}
