\name{NLmodel}
\docType{data}
\alias{NLmodel}
\title{The Nonlinear State-Space Model Data}
\description{The series generated by the nonlinear state-space model.}
\usage{
data(NLmodel)
}
\format{
A matrix with 100 rows and 2 columns.
\tabular{rl}{
  [, 1] \tab \eqn{x_n}{x(n)} \cr
  [, 2] \tab \eqn{y_n}{y(n)}
}
}
\details{
The system model \eqn{x_n}{x(n)} and the observation model \eqn{y_n}{y(n)} are
 generated by following state-space model:
\deqn{x_n = \frac{1}{2} x_{n-1} + \frac{25x_{n-1}}{x_{n-1}^2+1} + 8cos(1.2n) + v_n}{%
x(n) = 1/2 * x(n-1) + 25x(n-1) / (x(n-1)**2 + 1) + 8cos(1.2n) + v(n)}
\deqn{y_n = \frac{x_n^2}{10} + w_n,}{y(n) = x(n)**2 / 10 + w(n),}
where \eqn{v_n}{v(n)} ~ \eqn{N(0, 1)}, \eqn{w_n}{w(n)} ~ \eqn{N(0, 10)},
 \eqn{v_0}{v(0)} ~ \eqn{N(0, 5)}.
}
\keyword{datasets}
