\name{PfilterSample}
\docType{data}
\alias{PfilterSample}
\title{Sample Data for Particle Filter and Smoother}
\description{
An artificially generated sample data with shifting mean value.
}
\usage{
data(PfilterSample)
}
\format{
A time series of 400 observations.
}
\details{
This data generated by the following models;
  \tabular{rrlr}{
    \eqn{y_n} ~ \eqn{N(\mu_n, 1)}, \tab \eqn{\mu_n} \tab  = 0, \tab 1 <= n <= 100 \cr
    \tab \tab  = 1, \tab 101 <= n <= 200 \cr
    \tab \tab  = -1, \tab 201 <= n <= 300 \cr
    \tab \tab  = 0, \tab 301 <= n <= 400 
  }
}
\keyword{datasets}