\name{plot.arma}
\alias{plot.arma}
\title{Plot Analysis Result of ARMA Model}
\description{
  Plot impulse, autocovariance, PARCOR and characteristic roots of scalar
  ARMA model returned by \code{\link{armaimp}}.
}
\usage{
\method{plot}{arma}(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{"arma"} created by
           \code{\link{armaimp}}.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
