\name{plot.boxcox}
\alias{plot.boxcox}
\title{Plot Box-Cox Transformed Data}
\description{
  Plot original data and transformed data with minimum AIC.
}
\usage{
\method{plot}{boxcox}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"boxcox"}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
