\name{plot.lsqr}
\alias{plot.lsqr}
\title{Plot Fitted Trigonometric Polynomial}
\description{
  Plot original data and fitted trigonometric polynomial returned by
  \code{\link{lsqr}}.
}
\usage{
\method{plot}{lsqr}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"lsqr"}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
