\name{plot.polreg}
\alias{plot.polreg}
\title{Plot Fitted Polynomial Trend}
\description{
  Plot trend component of fitted polynomial returned by \code{\link{polreg}}.
}
\usage{
\method{plot}{polreg}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"polreg"}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
