\name{plot.season}
\alias{plot.season}
\title{Plot Trend, Seasonal and AR Components}
\description{
  Plot trend component, seasonal component, AR component and noise returned by
  \code{\link{season}}.
}
\usage{
\method{plot}{season}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"season"}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
