\name{plot.smooth}
\alias{plot.smooth}
\title{Plot Posterior Distribution of Smoother}
\description{
  Plot posterior distribution (mean and standard deviations) of the smoother returned by
  \code{\link{tsmooth}}.
}
\usage{
\method{plot}{smooth}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"smooth"}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
