\name{plot.spg}
\alias{plot.spg}
\title{Plot Smoothed Periodogram}
\description{
  Plot smoothed periodogram or logarithm of smoothed periodogram.
}
\usage{
\method{plot}{spg}(x, \dots)
}
\arguments{
  \item{x}{an object of class \code{"spg"} created by \code{\link{period}} and
    \code{\link{fftper}}.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
