\name{plot.trend}
\alias{plot.trend}
\title{Plot Trend and Residuals}
\description{
  Plot trend component and residuals returned by \code{\link{trend}}.
}
\usage{
\method{plot}{trend}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"trend"}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
