\name{Rainfall}
\alias{Rainfall}
\docType{data}
\title{Rainfall Data}
\description{
Number of rainy days in two years (1975-1976) at Tokyo, Japan.
}
\usage{
data(Rainfall)
}
\format{
Integer-valued time series of 366 observations.
}
\source{
The data were obtained from Tokyo District Meteorological Observatory.
\url{https://www.data.jma.go.jp/obd/stats/etrn/}
}
\keyword{datasets}
