\name{Sunspot}
\docType{data}
\alias{Sunspot}
\title{Sunspot Number Data}
\description{
  Yearly numbers of sunspots from to 1749 to 1979.
}
\usage{
data(Sunspot)
}
\format{
  A time series of 231 observations; yearly from 1749 to 1979.
}
\details{
  Sunspot is a part of the dataset \code{\link{sunspot.year}} from 1700 to 1988.
  Value "0" is converted into "0.1" for log transformation.
}
\keyword{datasets}