\name{Temperature}
\docType{data}
\alias{Temperature}
\title{Temperatures Data}
\description{
The daily maximum temperatures in Tokyo (from 1979-01-01 to 1980-04-30).
}
\usage{
data(Temperature)
}
\format{
A time series of 486 observations.
}
\source{
The data were obtained from Tokyo District Meteorological Observatory.
\url{https://www.data.jma.go.jp/obd/stats/etrn/}
}
\keyword{datasets}
 