\name{TSSS-package}
\alias{TSSS-package}
\alias{TSSS}
\docType{package}

\title{
  Time Series Analysis with State Space Model
}
\description{
  R functions for statistical analysis, modeling and simulation of time series
  with state space model.
}
\details{
  This package provides functions for statistical analysis, modeling and
  simulation of time series. These functions are developed based on source code
  of "FORTRAN 77 Programming for Time Series Analysis".

  After that, the revised edition "Introduction to Time Series Analysis (in Japanese)" 
  and the translation version "Introduction to Time Series Modeling" are published.

  Currently the revised edition "Introduction to Time Series Modeling with Applications in R"
  is published, in which calculations of most of the modeling or methods are explained using
  this package.
}
\references{
  Kitagawa, G. and Gersch, W. (1996)
  \emph{Smoothness Priors Analysis of Time Series}. Lecture Notes in Statistics,
  No.116, Springer-Verlag.

  Kitagawa, G. (2010)
  \emph{Introduction to Time Series Modeling}. Chapman & Hall/CRC.

  Kitagawa, G. (2020)
  \emph{Introduction to Time Series Modeling with Applications in R}. Chapman & Hall/CRC.

  Kitagawa, G. (1993)
  \emph{FORTRAN 77 Programming for Time Series Analysis}.
  The Iwanami Computer Science Series, Iwanami Publishing Company (in Japanese).

  Kitagawa, G. (2005)
  \emph{Introduction to Time Series Analysis}.
  Iwanami Publishing Company (in Japanese).

  Kitagawa, G. (2020)
  \emph{Introduction to Time Series Modeling with R}.
  Iwanami Publishing Company (in Japanese).
}

\keyword{package}


