\name{WHARD}
\docType{data}
\alias{WHARD}
\title{Wholesale Hardware Data}
\description{
The monthly record of wholesale hardware data. (January 1967 - November 1979)
}
\usage{
data(WHARD)
}
\format{
A time series of 155 observations.
}
\source{
The data were obtained from the United States Bureau of Labor Statistics (BLS).
}
\keyword{datasets}