\name{plot.smooth}
\alias{plot.smooth}
\title{Plot Mean Vectors of Smoother}
\description{
  Plot Mean vectors of the smoother and standard deviation returned by
  \code{\link{tsmooth}}.
}
\usage{
\method{plot}{smooth}(x, rdata = NULL, \dots)
}
\arguments{
  \item{x}{an object of class \code{"smooth"} created by \code{\link{tsmooth}}.}
  \item{rdata}{original data, if necessary.}
  \item{\dots}{further graphical parameters may also be supplied as arguments.}
}
\keyword{ts}
