\name{TSSS-package}
\alias{TSSS-package}
\alias{TSSS}
\docType{package}

\title{
  Time Series Analysis with State Space Model
}
\description{
  R functions for statistical analysis, modeling and simulation of time series
  with state space model.
}
\details{
  This package provides functions for statistical analysis, modeling and
  simulation of time series. These functions are developed based on source code
  of "FORTRAN 77 Programming for Time Series Analysis".

  Now, a revised edition "Introduction to Time Series Analysis (in Japanese)" 
  and "Introduction to Time Series Modeling" are published.
}
\references{
  Kitagawa, G. (2010)
  \emph{Introduction to Time Series Modeling}. Chapman & Hall/CRC.

  Kitagawa, G. and Gersch, W. (1996)
  \emph{Smoothness Priors Analysis of Time Series}. Lecture Notes in Statistics,
  No.116, Springer-Verlag.

  Kitagawa, G. (2005)
  \emph{Introduction to Time Series Analysis (in Japanese)}.
  Iwanami Publishing Company.

  Kitagawa, G. (1993)
  \emph{FORTRAN 77 Programming for Time Series Analysis (in Japanese)}.
  The Iwanami Computer Science Series.
}

\keyword{package}


